+++ backend/models/response.go
@@ -27,7 +27,7 @@ type OrderListResponse struct {
 	ShopName     string       `json:"shop_name"`
 	Location     string       `json:"location"`
 	OrderDate    time.Time    `json:"order_date"`
-	TotalAmount  float64      `json:"total_amount"`
+	TotalAmount  int          `json:"total_amount"`
 	Status       string       `json:"status"` // "cooking" or "completed"
 	WaitingCount int          `json:"waiting_count"`
 	Items        []ItemDetail `json:"items"`
@@ -58,7 +58,7 @@ type AdminOrderResponse struct {
 	OrderID       int          `json:"order_id"`
 	CustomerEmail *string      `json:"customer_email"`
 	OrderDate     time.Time    `json:"order_date"`
-	TotalAmount   float64      `json:"total_amount"`
+	TotalAmount   int          `json:"total_amount"`
 	Status        string       `json:"status"`
 	Items         []ItemDetail `json:"items"`
 }
